等待接单倒计时marker

+++ android/src/main/kotlin/me/yohom/amapbase/map/marker/SimpleLoc.java
@@ -3,7 +3,7 @@ package me.yohom.amapbase.map.marker;
 public class SimpleLoc {
     public double lat;
     public double lng;
-    public double bearing;
+    public Float bearing;
 
     @Override
     public String toString() {
